# docs/swagger/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the FeatherWings daemon.",
        "title": "FeatherWings API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/servers/{server}": {
            "get": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Get server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server identifier",
                        "name": "server",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Delete server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server identifier",
                        "name": "server",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/servers/{server}/commands": {
            "post": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Send console commands",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server identifier",
                        "name": "server",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Commands",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router.ServerCommandsRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/servers/{server}/deleteAllBackups": {
            "delete": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Delete all server backups",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server identifier",
                        "name": "server",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/servers/{server}/install": {
            "post": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Trigger server install",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server identifier",
                        "name": "server",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/servers/{server}/install-logs": {
            "get": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Retrieve server install logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server identifier",
                        "name": "server",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/router.ServerInstallLogResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/servers/{server}/logs": {
            "get": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Tail server logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server identifier",
                        "name": "server",
                        "in": "path",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Number of lines",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/router.ServerLogResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/servers/{server}/power": {
            "post": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Execute server power action",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server identifier",
                        "name": "server",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Power action",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router.ServerPowerRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/servers/{server}/reinstall": {
            "post": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Reinstall server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server identifier",
                        "name": "server",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/servers/{server}/sync": {
            "post": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Sync server state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server identifier",
                        "name": "server",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/servers/{server}/ws/deny": {
            "post": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Invalidate websocket tokens",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server identifier",
                        "name": "server",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tokens",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router.ServerDenyTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "environment.Variables": {
            "type": "object",
            "additionalProperties": true
        },
...
        "router.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "router.ServerCommandsRequest": {
            "type": "object",
            "properties": {
                "commands": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "router.ServerDenyTokenRequest": {
            "type": "object",
            "properties": {
                "jtis": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "router.ServerInstallLogResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                }
            }
        },
        "router.ServerLogResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "router.ServerPowerRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "wait_seconds": {
                    "type": "integer"
                }
            }
        },
...
        "server.APIResponse": {
            "type": "object",
            "properties": {
                "configuration": {
                    "$ref": "#/definitions/server.Configuration"
                },
                "is_suspended": {
                    "type": "boolean"
                },
                "state": {
                    "type": "string"
                },
                "utilization": {
                    "$ref": "#/definitions/server.ResourceUsage"
                }
            }
        },
        "server.Configuration": {
            "type": "object",
            "properties": {
                "allocations": {
                    "$ref": "#/definitions/environment.Allocations"
                },
                "build": {
                    "$ref": "#/definitions/environment.Limits"
                },
                "container": {
                    "type": "object",
                    "properties": {
                        "image": {
                            "description": "Defines the Docker image that will be used for this server",
                            "type": "string"
                        }
                    }
                },
                "crash_detection_enabled": {
                    "type": "boolean"
                },
                "egg": {
                    "$ref": "#/definitions/server.EggConfiguration"
                },
                "environment": {
                    "description": "An array of environment variables that should be passed along to the running\nserver process.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/environment.Variables"
                        }
                    ]
                },
                "invocation": {
                    "description": "The command that should be used when booting up the server instance.",
                    "type": "string"
                },
                "labels": {
                    "description": "Labels is a map of container labels that should be applied to the running server process.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/server.ConfigurationMeta"
                },
                "mounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.Mount"
                    }
                },
                "skip_egg_scripts": {
                    "description": "By default this is false, however if selected within the Panel while installing or re-installing a\nserver, specific installation scripts will be skipped for the server process.",
                    "type": "boolean"
                },
                "suspended": {
                    "description": "Whether or not the server is in a suspended state. Suspended servers cannot\nbe started or modified except in certain scenarios by an admin user.",
                    "type": "boolean"
                },
                "uuid": {
                    "description": "The unique identifier for the server that should be used when referencing\nit against the Panel API (and internally). This will be used when naming\ndocker containers as well as in log output.",
                    "type": "string"
                }
            }
        },
        "server.ConfigurationMeta": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "NodeToken": {
            "description": "Supply the node's bearer token from `config.yml` using the `Authorization: Bearer \u003ctoken\u003e` header.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ServerJWT": {
            "description": "Signed JWTs issued by the Panel for server-scoped operations (uploads, downloads, websockets). Pass in the `token` query parameter.",
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}